Reorganized CSS into one file. Made poll voting more accessible.
