--- src/CreatePoll.css
@@ -1,62 +1,0 @@
-
-.create {
-  width: 80%;
-  max-width: 600px;
-  margin: 0 auto;
-  /* height: 100%; */
-  display: flex;
-  /* flex: 1 0 auto; */
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-}
-
-.create__form {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-}
-
-.create__button {
-  margin-top: 1rem;
-}
-
-.create__slug {
-  display: flex;
-  align-items: center;
-}
-
-.create__slug__address {
-  font-size: 1rem;
-}
-.create__slug__input {
-  width: 6rem;
-  height: 2rem;
-  padding: .5rem;
-  font-size: 1rem;
-  margin-left: 1rem;
-}
-
-.create__slug__input--taken {
-  border: 1px solid red;
-}
-
-.poll-options {
-  font-size: 1.5rem;
-  width: 100%;
-}
-
-.poll-option {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  width: 100%;
-  margin: 0 auto;
-  align-self: center
-}
-
-.poll-option__input {
-  width: 80%;
-  margin: .2rem;
-}
