--- src/Results.css
@@ -1,203 +1,0 @@
-.results {
-  position: relative;
-  width: 80%;
-  margin: 0 auto;
-  max-width: 800px;
-  text-align: center;
-}
-
-.results__list {
-  position: relative;
-  padding: 1rem 0 0.5rem;
-  width: 100%;
-  overflow: hidden;
-}
-
-.results__list::before {
-  content: '';
-  bottom: 0.5rem;
-  left: 0;
-  width: 16px;
-  height: 100%;
-  background-color: hsl(0, 0%, 80%);;
-  position: absolute;
-  z-index: -1;
-  transform: skewY(-51deg);
-  /* opacity: .8; */
-}
-
-.results__list::after {
-  content: '';
-  bottom: -0.1rem;
-  left: 0.34rem;
-  width: 101%;
-  height: 1.1rem;
-  transform: skewX(-39deg);
-  background-color: hsl(0, 0%, 40%);;
-  position: absolute;
-  z-index: -1;
-}
-
-.results__list__percentage {
-  color: white;
-  font-size: 1.1rem;
-  position: absolute;
-  bottom: .3rem;
-  transform: translateX(-50%) skewX(15deg) skewY(-25deg);
-  font-family: 'Fredericka the Great', cursive;
-}
-
-.results__list__percentage--25 {
-  left: 25%;
-}
-
-.results__list__percentage--50 {
-  left: 50%;
-}
-
-.results__list__percentage--75 {
-  left: 75%;
-}
-
-.results__list__vote-count {
-  position: absolute;
-  top: .3rem;
-  right: .5rem;
-  content: '16 votes';
-  color: white;
-  text-shadow: -1px 1px hsl(0, 0%, 10%);
-  background-color: hsl(0, 0%, 20%);
-}
-
-.results__list__background {
-  position: absolute;
-  bottom: 1rem;
-  left: .8rem;
-  width: 100%;
-  height: 100%;
-  background-color: hsl(0, 0%, 20%);
-  z-index: -10;
-}
-
-.result {
-  position: relative;
-  display: flex;
-  align-items: center;
-  margin: 2rem 0;
-  height: 3rem;
-  width: 100%;
-  border-top-right-radius: 5px;
-  border-bottom-right-radius: 5px;
-}
-
-.result__title {
-  color: #eee;
-  font-size: 1.8rem;
-  padding-left: 1rem;
-  font-family: 'Fredericka the Great', cursive;
-}
-
-.result__title--light {
-  color: #eee;
-}
-
-.result__bar {
-  position: absolute;
-  display: flex;
-  align-items: center;
-  top: 0;
-  background: hsl(283, 55%, 30%);
-  color: #eee;
-  height: 100%;
-  animation: 1s grow linear forwards;
-  white-space: nowrap;
-}
-
-.result__bar__flag {
-  content: attr(votes);
-  position: absolute;
-  display: grid;
-  top: -1rem;
-  left: calc(100% + .7rem);
-  align-items: center;
-  padding: 0 1rem 0 .5rem;
-  background: rgb(243, 46, 46);
-  width: fit-content;
-  height: 1.5rem;
-  color: white;
-  z-index: -1;
-}
-
-.result__bar__flag::after {
-  content: '';
-  position: absolute;
-  top: .3rem;
-  left: 85%;
-  width: 1.5rem;
-  height: 1.5rem;
-  background-color: hsla(0, 0%, 20%);
-  transform: rotate(-67deg) skewX(-45deg);
-  z-index: -1;
-}
-
-.result__bar__shadow {
-  height: 3rem;
-  background: hsl(283, 100%, 40%);
-  position: absolute;
-  left: .8rem;
-  bottom: 1rem;
-  z-index: -1;
-  animation: 1s grow linear forwards;
-}
-
-.result__bar__shadow::before {
-  content: '';
-  position: absolute;
-  top: .27rem;
-  left: -.55rem;
-  width: 1.3rem;
-  height: 1.3rem;
-  background: linear-gradient(51deg, hsl(283, 100%, 40%) 0 47%, transparent 47% 100%);
-  transform: rotate(39deg);
-}
-
-.result__bar__shadow::after {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  content: '';
-  font-size: 10px;
-  position: absolute;
-  top: 10px;
-  left: calc(100% - 16px);
-  background: hsl(283, 100%, 70%);;
-  width: .81rem;
-  height: 3rem;
-  transform: skewY(-51deg);
-}
-
-.results__share {
-  margin-top: 1rem;
-  display: flex;
-  width: 100%;
-  justify-content: space-evenly;
-}
-
-.results__share__button {
-  display: flex;
-  align-items: center;
-  justify-content: space-around;
-  font-size: 1.2rem;
-  width: 10rem;
-  height: 2.5rem;
-}
-
-.results__share__button i {
-  font-size: 1rem;
-}
-
-@keyframes grow {
-  0% {
-    width: 0%;
-  }
-}
