+++ src/styles/App.css
@@ -215,6 +215,360 @@ footer {
   border-radius: 2px 2px 25px 25px;
 }
 
+
+/* ================= */
+/* CreatePoll Styles */
+/* ================= */
+
+
+.create {
+  width: 80%;
+  max-width: 600px;
+  margin: 0 auto;
+  /* height: 100%; */
+  display: flex;
+  /* flex: 1 0 auto; */
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+}
+
+.create__form {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+}
+
+.create__button {
+  margin-top: 1rem;
+}
+
+.create__slug {
+  display: flex;
+  align-items: center;
+}
+
+.create__slug__address {
+  font-size: 1rem;
+}
+.create__slug__input {
+  width: 6rem;
+  height: 2rem;
+  padding: .5rem;
+  font-size: 1rem;
+  margin-left: 1rem;
+}
+
+.create__slug__input--taken {
+  border: 1px solid red;
+}
+
+.poll-options {
+  font-size: 1.5rem;
+  width: 100%;
+}
+
+.poll-option {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  width: 100%;
+  margin: 0 auto;
+  align-self: center
+}
+
+.poll-option__input {
+  width: 80%;
+  margin: .2rem;
+}
+
+/* ============== */
+/* Results styles */
+/* ============== */
+
+.results {
+  position: relative;
+  width: 80%;
+  margin: 0 auto;
+  max-width: 800px;
+  text-align: center;
+}
+
+.results__list {
+  position: relative;
+  padding: 1rem 0 0.5rem;
+  width: 100%;
+  overflow: hidden;
+}
+
+.results__list::before {
+  content: '';
+  bottom: 0.5rem;
+  left: 0;
+  width: 16px;
+  height: 100%;
+  background-color: hsl(0, 0%, 80%);;
+  position: absolute;
+  z-index: -1;
+  transform: skewY(-51deg);
+  /* opacity: .8; */
+}
+
+.results__list::after {
+  content: '';
+  bottom: -0.1rem;
+  left: 0.34rem;
+  width: 101%;
+  height: 1.1rem;
+  transform: skewX(-39deg);
+  background-color: hsl(0, 0%, 40%);;
+  position: absolute;
+  z-index: -1;
+}
+
+.results__list__percentage {
+  color: white;
+  font-size: 1.1rem;
+  position: absolute;
+  bottom: .3rem;
+  transform: translateX(-50%) skewX(15deg) skewY(-25deg);
+  font-family: 'Fredericka the Great', cursive;
+}
+
+.results__list__percentage--25 {
+  left: 25%;
+}
+
+.results__list__percentage--50 {
+  left: 50%;
+}
+
+.results__list__percentage--75 {
+  left: 75%;
+}
+
+.results__list__vote-count {
+  position: absolute;
+  top: .3rem;
+  right: .5rem;
+  content: '16 votes';
+  color: white;
+  text-shadow: -1px 1px hsl(0, 0%, 10%);
+  background-color: hsl(0, 0%, 20%);
+}
+
+.results__list__background {
+  position: absolute;
+  bottom: 1rem;
+  left: .8rem;
+  width: 100%;
+  height: 100%;
+  background-color: hsl(0, 0%, 20%);
+  z-index: -10;
+}
+
+.result {
+  position: relative;
+  display: flex;
+  align-items: center;
+  margin: 2rem 0;
+  height: 3rem;
+  width: 100%;
+  border-top-right-radius: 5px;
+  border-bottom-right-radius: 5px;
+}
+
+.result__title {
+  color: #eee;
+  font-size: 1.8rem;
+  padding-left: 1rem;
+  font-family: 'Fredericka the Great', cursive;
+}
+
+.result__title--light {
+  color: #eee;
+}
+
+.result__bar {
+  position: absolute;
+  display: flex;
+  align-items: center;
+  top: 0;
+  background: hsl(283, 55%, 30%);
+  color: #eee;
+  height: 100%;
+  animation: 1s grow linear forwards;
+  white-space: nowrap;
+}
+
+.result__bar__flag {
+  content: attr(votes);
+  position: absolute;
+  display: grid;
+  top: -1rem;
+  left: calc(100% + .7rem);
+  align-items: center;
+  padding: 0 1rem 0 .5rem;
+  background: rgb(243, 46, 46);
+  width: fit-content;
+  height: 1.5rem;
+  color: white;
+  z-index: -1;
+}
+
+.result__bar__flag::after {
+  content: '';
+  position: absolute;
+  top: .3rem;
+  left: 85%;
+  width: 1.5rem;
+  height: 1.5rem;
+  background-color: hsla(0, 0%, 20%);
+  transform: rotate(-67deg) skewX(-45deg);
+  z-index: -1;
+}
+
+.result__bar__shadow {
+  height: 3rem;
+  background: hsl(283, 100%, 40%);
+  position: absolute;
+  left: .8rem;
+  bottom: 1rem;
+  z-index: -1;
+  animation: 1s grow linear forwards;
+}
+
+.result__bar__shadow::before {
+  content: '';
+  position: absolute;
+  top: .27rem;
+  left: -.55rem;
+  width: 1.3rem;
+  height: 1.3rem;
+  background: linear-gradient(51deg, hsl(283, 100%, 40%) 0 47%, transparent 47% 100%);
+  transform: rotate(39deg);
+}
+
+.result__bar__shadow::after {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  content: '';
+  font-size: 10px;
+  position: absolute;
+  top: 10px;
+  left: calc(100% - 16px);
+  background: hsl(283, 100%, 70%);;
+  width: .81rem;
+  height: 3rem;
+  transform: skewY(-51deg);
+}
+
+.results__share {
+  margin-top: 1rem;
+  display: flex;
+  width: 100%;
+  justify-content: space-evenly;
+}
+
+.results__share__button {
+  display: flex;
+  align-items: center;
+  justify-content: space-around;
+  font-size: 1.2rem;
+  width: 10rem;
+  height: 2.5rem;
+}
+
+.results__share__button i {
+  font-size: 1rem;
+}
+
+@keyframes grow {
+  0% {
+    width: 0%;
+  }
+}
+
+/* Voting Styles */
+
+.poll {
+  width: 80%;
+  margin: 0 auto;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+}
+
+.poll__choice {
+  width: 100%;
+  max-width: 600px;
+}
+
+.poll__choice__label {
+  margin-bottom: 1rem;
+  padding: .5rem;
+  display: block;
+  border: none;
+  font-family: 'Fredericka the Great', sans-serif;
+  font-size: 1.5rem;
+  text-align: center;
+  padding: 1rem 1.5rem;
+  border-radius: 100vmax;
+  background-color: hsl(0, 0%, 24%);
+  box-shadow: .2rem .2rem .3rem rgba(12, 12, 12),
+  inset -.2rem -.2rem .3rem hsl(0, 0%, 10%),
+  inset .2rem .2rem .3rem hsl(0, 0%, 80%);
+  transition: background-color .3s ease-out, box-shadow .3s ease-out, transform .3s ease-out;
+  cursor: pointer;
+  color: white;
+  word-break: break-all;
+}
+
+.poll__choice__label:hover {
+  background-color: hsl(0, 0%, 24%);
+  box-shadow: .3rem .3rem .4rem rgba(12, 12, 12),
+              inset -.4rem -.4rem .6rem hsl(0, 0%, 10%),
+              inset .4rem .4rem .6rem hsl(0, 0%, 80%);
+}
+
+.poll__choice__label:focus {
+  box-shadow: 0 0 0 .25rem hsl(0, 0%, 95%),
+              0 0 0 .5rem hsl(0, 0%, 30%),
+              inset -.2rem -.2rem .3rem hsl(0, 0%, 10%),
+              inset .2rem .2rem .3rem hsl(0, 0%, 80%);
+              outline: none;
+}
+
+
+[type="radio"]:checked ~ .poll__choice__label,
+.poll__choice__label:active {
+  box-shadow: .2rem .2rem .3rem rgba(12, 12, 12),
+              inset -.2rem -.2rem .3rem hsl(0, 0%, 10%),
+              inset .2rem .2rem .3rem hsl(0, 0%, 80%),
+              inset .4rem .4rem .6rem hsl(0, 0%, 10%),
+              inset -.4rem -.4rem .6rem hsl(0, 0%, 80%);
+}
+
+.poll__button {
+  margin-top: 3rem;
+  width: 70%;
+  max-width: 500px;
+  transition: opacity .3s ease-out, background-color .1s ease-out, box-shadow .1s ease-out, transform .1s ease-out;
+}
+
+.poll__button--hidden {
+  opacity: 0;
+  cursor: initial;
+  transform: translateY(2rem);
+}
+
+/* ============= */
+/* Footer styles */
+/* ============= */
+
 footer p, footer span {
   font-size: clamp(1rem, 5vw, 1.5rem);
 }
